cleanup: Refactor the code about configuration
