+++ kubespider/api/types.py
@@ -1,3 +1,5 @@
+# Used to define the api types in the project
+
 SOURCE_PROVIDER_PERIOD_TYPE = "period"
 SOURCE_PROVIDER_DISPOSABLE_TYPE = "disposable"
 
@@ -11,14 +13,3 @@ FILE_TYPE_VIDEO_MOVIE = "movie"
 FILE_TYPE_VIDEO_MIXED = "video_mixed"
 FILE_TYPE_MUSIC = "music"
 FILE_TYPE_PICTIRE = "picture"
-
-file_type_to_path = {
-    FILE_TYPE_COMMON: "Common",
-    FILE_TYPE_VIDEO_TV: "TV",
-    FILE_TYPE_VIDEO_MOVIE: "Movie",
-    FILE_TYPE_VIDEO_MIXED: "VideoMixed",
-    FILE_TYPE_MUSIC: "Music",
-    FILE_TYPE_PICTIRE: "Picture"
-}
-
-PT_BOT_PATH = "PT"
