+++ kubespider/utils/helper.py
@@ -1,29 +1,15 @@
-import os
 import uuid
 import hashlib
 import logging
 import cgi
 import urllib
-from enum import Enum
 from urllib.parse import urlparse
 from urllib import request
-from utils.config_reader import YamlFileConfigReader
+
 from api import types
+from api import values
+from utils import global_config
 
-class Config(str, Enum):
-    SOURCE_PROVIDER = 'source_provider.yaml'
-    DOWNLOAD_PROVIDER = 'download_provider.yaml'
-    PT_PROVIDER = 'pt_provider.yaml'
-    KUBESPIDER_CONFIG = 'kubespider.yaml'
-    STATE = 'state.yaml'
-
-    def __str__(self) -> str:
-        return str(self.value)
-
-    def config_path(self) -> str:
-        return os.path.join(cfg_base_path, self)
-
-cfg_base_path = config_path = os.path.join(os.getenv('HOME'), '.config/')
 
 def get_tmp_file_name(url):
     file_name = get_unique_hash(url)
@@ -36,8 +22,8 @@ def get_unique_hash(data):
     return hashlib.md5(data.encode('utf-8')).hexdigest()
 
 def convert_file_type_to_path(file_type: str):
-    if file_type in types.file_type_to_path.keys():
-        return types.file_type_to_path[file_type]
+    if file_type in values.FILE_TYPE_TO_PATH.keys():
+        return values.FILE_TYPE_TO_PATH[file_type]
     logging.warning('%s file file is not recorded', file_type)
     return file_type
 
@@ -46,16 +32,8 @@ def format_long_string(longstr: str) -> str:
         return longstr[:40] + '...'
     return longstr
 
-global_config = YamlFileConfigReader(Config.KUBESPIDER_CONFIG.config_path())
-
-def get_proxy() -> str:
-    cfg = global_config.read()
-    if cfg is not None:
-        return cfg.get('proxy', None)
-    return None
-
 def get_request_controller() -> request.OpenerDirector:
-    proxy_addr = get_proxy()
+    proxy_addr = global_config.get_proxy()
 
     proxy_handler = None
     handler: request.OpenerDirector = None
@@ -96,9 +74,3 @@ def parse_cookie_string(cookie: str) -> dict:
         key, value = item.strip().split('=')
         cookie_dict[key] = value
     return cookie_dict
-
-def get_auth_token() -> str:
-    cfg = global_config.read()
-    if cfg is not None:
-        return cfg.get('auth_token', None)
-    return None
